[JENKINS-53858] Remove synchronized from connect() to fix deadlock
There are multiple codepaths that can call connect(). Some codepaths
have already acquired the hudson.model.Queue lock, and others don't, so
the synchronized keyword here can lead to deadlocks where one thread is
holding the Queue lock and trying to lock the EC2Cloud, and another
thread is holding the EC2Cloud lock and trying to lock the Queue.

With the keyword removed, there's the possibility of multiple EC2
connections occurring, but I've yet to see this cause any observable
problems in reality, and the deadlock issue seems to have stopped
entirely.